+++ topalias/__init__.py
@@ -10,11 +10,7 @@ import os
 import sys
 
 import requests
-
-try:
-    from packaging.version import parse
-except ImportError:
-    from pip._vendor.packaging.version import parse
+from packaging.version import parse  # noqa: WPS347
 
 sys.path.insert(
     0,
@@ -28,12 +24,14 @@ sys.path.insert(
 
 TOPALIAS_PYPI_LATEST_VERSION = "https://pypi.python.org/pypi/topalias/json"
 
+TOPALIAS_EXAMPLES = "https://raw.githubusercontent.com/CSRedRat/topalias/master/topalias/data/.bash_aliases"
+
 
 def get_version(url=TOPALIAS_PYPI_LATEST_VERSION):
     """Return version of topalias package on pypi.org."""
     req = requests.get(url)
     version = parse("0")
-    if req.status_code == requests.codes.ok:
+    if req.status_code == requests.codes.ok:  # pylint: disable=no-member
         req.encoding = req.apparent_encoding
         j = json.loads(req.text.encode(req.encoding))
         releases = j.get("releases", [])
@@ -42,3 +40,13 @@ def get_version(url=TOPALIAS_PYPI_LATEST_VERSION):
             if not ver.is_prerelease:
                 version = max(version, ver)
     return version
+
+
+def get_examples(url=TOPALIAS_EXAMPLES):
+    """Give dynamic .bash_aliases example from GitHub."""
+    req = requests.get(url)
+    alias_examples = ""
+    if req.status_code == requests.codes.ok:  # pylint: disable=no-member
+        req.encoding = req.apparent_encoding
+        alias_examples = req.text
+    return alias_examples