+++ topalias/aliascore.py
@@ -6,19 +6,22 @@ import os
 import random
 import re
 import sys
-from collections import defaultdict
 
-import statistic
+from statistic import most_used_utils, top_command
 
 NOTHING = "Empty"
 
-home = os.path.expanduser("~")
-path = [".", home]
-suggestion_count = 20
-debug = False
+DEBUG = False
+HOME = os.path.expanduser("~")
+path = [".", HOME]
+SUGGESTION_COUNT = 20
+ALIASES_FILTER = False
+HISTORY_FILE = ".bash_history"
 
 logging.basicConfig(
-    stream=sys.stdout, format='%(levelname)s:%(message)s', level=(logging.DEBUG if debug else logging.ERROR)
+    stream=sys.stdout,
+    format="%(levelname)s:%(message)s",
+    level=(logging.DEBUG if DEBUG else logging.ERROR),
 )
 
 
@@ -39,13 +42,14 @@ def find_first(filename: str, paths: list) -> str:  # type: ignore
 
 def find_history() -> str:
     """Find command history file"""
-    history_name: str = ".bash_history"
-    history_path = find_first(history_name, path)
+    history_path = find_first(HISTORY_FILE, path)
     if history_path != NOTHING:
+        logging.debug("History file: %s", history_path)
         return history_path
-    print("File {} not found in any of the directories".format(history_name))
+    print("File {} not found in any of the directories".format(HISTORY_FILE))
     file_dir = os.path.dirname(os.path.realpath("__file__"))
-    data_path = os.path.join(file_dir, r"topalias/data/.bash_history")
+    data_path = os.path.join(file_dir, r"topalias/data/" + HISTORY_FILE)
+    logging.debug("History file: %s", data_path)
     return data_path  # noqa: WPS331
 
 
@@ -61,35 +65,19 @@ def find_aliases() -> str:
     return data_path  # noqa: WPS331
 
 
-def top_command(command, limit) -> list:  # type: ignore
-    """List top executed command from history"""
-    counts = defaultdict(int)  # type: ignore
-    for x in command:
-        counts[x] += 1
-    return sorted(counts.items(), reverse=True, key=lambda tup: tup[1])[
-           :limit
-           ]
+used_alias = []
 
 
-def top_alias():
+def collect_alias():
     """Top used aliases"""
 
-    alias_bank = []
-    with open(find_aliases(), "r", encoding="utf-8") as f:
-        for line in f:
-            if not line.startswith("#", 0, 1):
+    with open(find_aliases(), "r", encoding="utf-8") as aliases_data:
+        for line in aliases_data:
+            if not line.startswith("#", 0, 1) or line:
                 s = line.rstrip()
-                alias_name = list(s.split())
-                try:
-                    if alias_name[0] == "alias":
-                        print(alias_name[1])
-                except:
-                    pass
-                alias_bank.append(s)
-
-    print("Top used aliases: ")
-    print("For suggest new amazing short aliases run:\ntopalias h")
-    print("Print help:\ntopalias -h")
+                alias_name = list(s.split(" "))
+                if alias_name[0] == "alias":
+                    used_alias.append(alias_name[1].split("=")[0])
 
 
 acronyminator = re.compile(r"(?:(?<=\s)|^)(?:[a-z]|\d+)")
@@ -108,20 +96,9 @@ def filter_alias_length(raw_command_bank, min_length: int) -> list:  # type: ign
         if len(gen_alias) >= min_length:
             filtered_bank.append(command)
         else:
-            logging.info(f"COMMAND_FILTERED: {command}")
+            logging.info("COMMAND_FILTERED: %s", command)
 
     return filtered_bank
-
-
-def most_used_utils(raw_command_bank, limit=5) -> list:  # type: ignore
-    """Return first utility from command usage statistic"""
-    utility_bank = []
-    for command in raw_command_bank:
-        utility = command.split()[0]
-        utility_bank.append(utility)
-
-    utility_rank = top_command(utility_bank, limit)
-    return utility_rank
 
 
 def print_stat(raw_lines, filtered) -> None:
@@ -129,10 +106,25 @@ def print_stat(raw_lines, filtered) -> None:
     rows_count = len(raw_lines)
     unique_count = len(set(raw_lines))
     filtered_count = unique_count - len(set(filtered))
+    top_utils = most_used_utils(load_command_bank(filtering=ALIASES_FILTER))
+    top_utils_text_line = ""
+    for paired_rank in top_utils:  # noqa: WPS440, WPS519
+        top_utils_text_line += f"{paired_rank[0]}: {paired_rank[1]}, "  # noqa: WPS441
+    top_utils_text_line = top_utils_text_line[:-2]
     print(
         f"\ncommands in history: {rows_count}, unique commands: {unique_count}, filtered by length: {filtered_count}\n",
-        f"most used utils: {', '.join(map(str, most_used_utils(load_command_bank())))}"  # TODO: need format statistic
+        f"most used utils: {top_utils_text_line}",
     )
+    if used_alias:
+        top_aliases = most_used_utils(load_command_bank(), aliases=used_alias)
+        top_aliases_text_line = ""
+        for paired_rank in top_aliases:  # noqa: WPS440, WPS519
+            top_aliases_text_line += (
+                f"{paired_rank[0]}: {paired_rank[1]}, "  # noqa: WPS441
+            )
+        top_aliases_text_line = top_aliases_text_line[:-2]
+        if top_aliases:
+            print(f" most used aliases: {top_aliases_text_line}")
 
 
 HISTTIMEFORMAT_FIRST = "Hint: add timestamps in history log: "
@@ -145,9 +137,10 @@ hint_bank = (
     HISTTIMEFORMAT,
     "Hint: add space ' ' before sensitive command in terminal for skip save current command in history!",
     "Hint: command 'sudo !!' after you forget add sudo before command in previous command",
-    "Hint: command !<history command number> for repeat command from history",
+    "Hint: command !<command number in history> for repeat command from history",
     "Hint: ignore command in history: echo \"export HISTIGNORE='ls -l:pwd:date:ll:ls:'\" >> ~/.bashrc",
     'Hint: ignore duplicates in history: echo "export HISTCONTROL=ignoreboth" >> ~/.bashrc',
+    "Hint: run 'alias' command to print all used aliases",
 )
 
 
@@ -157,32 +150,51 @@ def print_hint() -> None:
     print(random.choice(hint_bank))
 
 
-def load_command_bank():
+def load_command_bank(filtering=False):
+    """Read and parse shell command history file"""
     command_bank = []
-    with open(find_history(), "r", encoding="utf-8") as f:
-        for line in f:
-            if not line.startswith("#", 0, 1):
-                s = line.rstrip()
-                command_bank.append(s)
-            # l.append(line.split('#')[0].split())  # noqa: E800
+    with open(find_history(), "r", encoding="utf-8") as history_data:
+        for line in history_data:
+            if HISTORY_FILE == ".bash_history":
+                if not line.startswith("#", 0, 1):
+                    clear_line = line.rstrip()
+                    if filtering:
+                        first_word_in_command = clear_line.split()[0]
+                        if first_word_in_command not in used_alias:
+                            command_bank.append(clear_line)
+                    else:
+                        command_bank.append(clear_line)
+            else:
+                clear_line = line.split(";")[1].rstrip()
+                if filtering and clear_line:
+                    first_word_in_command = clear_line.split()[0]
+                    if first_word_in_command not in used_alias:
+                        command_bank.append(clear_line)
+                elif clear_line:
+                    command_bank.append(clear_line)
     return command_bank
 
 
 def print_history(acronym_length) -> None:  # noqa: WPS210
     """Main function for print top commands and suggestions aliases"""
-    if debug:
+    if DEBUG:
         logging.getLogger().setLevel(logging.DEBUG)
 
-    command_bank = load_command_bank()
+    command_bank = load_command_bank(filtering=ALIASES_FILTER)
 
     filtered_alias_bank = filter_alias_length(command_bank, acronym_length)
-    top_raw_list = top_command(filtered_alias_bank, suggestion_count)
+    top_raw_list = top_command(filtered_alias_bank, SUGGESTION_COUNT)
     print("\n")
+
+    if HISTORY_FILE == ".zsh_history":
+        aliases_output = "~/.zshrc"
+    else:
+        aliases_output = "~/.bash_aliases"
 
     for num, ranked_command in reversed(list(enumerate(top_raw_list, start=1))):
         acronym = "".join(acronyminator.findall(ranked_command[0]))
         linux_add_alias = (
-            f"echo \"alias {acronym}='{ranked_command[0]}'\" >> ~/.bash_aliases"
+            f"echo \"alias {acronym}='{ranked_command[0]}'\" >> {aliases_output}"
         )
         print(
             f"{num}. {ranked_command[0]}\n",  # noqa: WPS221