+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 We follow [Semantic Versions](https://semver.org/).
 
 
+## Version 2.0.0
+
+- Add zsh shell support (-z/--zsh flag)
+- Add version check (version/--version)
+- --filter flag exclude used alias from command top chart
+- --debug mode
+- -f change dot files search path
+- -c count suggestion parameter
+- Add "example" command for give aliases example
+- Add aliases usage statistic
+
 ## Version 1.1.7
 
 - Build and publish working package